requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
faiss-cpu==1.9.0
sentence-transformers==3.3.1
numpy==1.26.4
pymongo==4.10.1
